+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * [FEATURE] Add option to use jump hashing to load balance requests to memcached #1554
 * [FEATURE] Add status page for HA tracker to distributors #1546
+* [CHANGE] Added check for retention period to be a multiple of periodic table duration #1564
 
 ## 0.1.0 / 2019-08-07
 
@@ -13,3 +14,4 @@
 * [FEATURE] You can specify "heap ballast" to reduce Go GC Churn #1489
 * [BUGFIX] HA Tracker no longer always makes a request to Consul/Etcd when a request is not from the active replica #1516
 * [BUGFIX] Queries are now correctly cancelled by the query-frontend #1508
+* [CHANGE] Added check for retention period to be a multiple of periodic table duration